requirements.txt
# Legal Document Generator - Backend Requirements
# Python 3.10+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# File Handling
python-multipart>=0.0.6
aiofiles>=23.2.0

# Production Server
gunicorn>=21.2.0

# Environment
python-dotenv>=1.0.0
